% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dairyplantdata-package.R
\docType{package}
\name{dairyplantdata-package}
\alias{dairyplantdata}
\alias{dairyplantdata-package}
\title{dairyplantdata: Provides information on dairy food from all the dairy organization in the United States}
\description{
This package was developed to facilitate a research project at the Bureau of Business Research. Provides information on dairy food from all the dairy organization in the United States.
}
\author{
\strong{Maintainer}: Bhabishya Neupane \email{bhabishya.neupane@huskers.unl.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
